Turn classic scripts into Js modules to perform TLA
Since ItemTagDic.json and TagIdDic,json need to be get before filling 
rack, the `Dictionary` and `Tags` variable need await. Top Level Await 
only exists in Js module.
