+++ index.html
@@ -2,8 +2,8 @@
 <head>
   <meta charset="utf-8">
   <link rel="stylesheet" href="./styles/Homepage.css">
-  <script src="./src/Essential.js"></script>
-  <script src="./src/Collection.js"></script>
+  <script type="module" src="./src/Essential.js"></script>
+  <script type="module" src="./src/Collection.js"></script>
   <title>Save House</title>
 </head>
 <body>
